spotify/requirements.txt
mlx
notion-client
pandas
streamlit
python-dotenv
spotipy